perf: pack AbsRank to the vendor's 2-byte layout
AbsRank carried two i32s where the vendor's absRankType is char rank +
signed char hand (dds.h). The 8192-entry per-deal rel table shrinks
from 3.93 MB to 960 KiB - it is randomly probed in make3, move
generation, and the quick/later-tricks pruning, and at 16 solver
threads the i32 layout put ~60 MB of hot tables on a 16 MiB L3.
rank (0..=14) and hand (-1..=3) are lossless in i8; all hot-path reads
widen with i32::from, the only narrowing is in the cold rel build.

+++ src/later_tricks.rs
@@ -32,7 +32,7 @@ const DDS_HANDS: usize = 4;
 #[inline]
 fn abs_rank(rel: &[RelRanks], aggr: u16, k: usize, suit: usize) -> (i32, i32) {
     let entry = rel[aggr as usize].abs_rank[k][suit];
-    (entry.rank, entry.hand)
+    (i32::from(entry.rank), i32::from(entry.hand))
 }
 
 /// MIN-side end-game pruning. Mirrors the vendor's `LaterTricksMIN`.
